+++ CHANGELOG.md
@@ -14,6 +14,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [TO BE DEPRECATED]
 - apis "config" and "clean" methods. Mercury "instances" methods should be used instead.
 
+## [1.6.0] - 2020-01-11
+### Fixed
+- Upgrade axios version
+- Maintain trailing slash in urls with params
+
+### Changed
+- Upgrade path-to-regex dependency
+- Upgrade devDependencies
+
 ## [1.5.0] - 2019-11-24
 ### Changed
 - Upgrade @data-provider/core dependency